refactoring following #29
